Refactor Some Codes
